refactor(agents): the staging bodies move into their packages
Phase 4b complete. `host-stage.ts` was 813 lines with claude's and codex's
stagers inside it; what remains is the part with no agent in it — the generic
`stageAgentStaticForUpload`, the shared `~/.agents` stager, and the rsync/tar
primitives, which are exported now so an agent package can build its own stager
without re-implementing the plumbing or the cleanup contract.

Moved with them, because each had exactly one consumer and it was the stager:
`claude-hooks-filter.ts` -> `agent-claude/src/hooks-filter.ts`, `codex-config.ts`
-> `agent-codex/src/box-config.ts`, and `claude-json-overlay.ts` ->
`agent-claude/src/cloud-json-overlay.ts`. Allowlist 21 -> 19.

The overlay forced the item this plan had deferred: `seedClaudeJsonAtCreate` was
still called BY NAME from `cloud-sync.ts`, and once it moved, `sandbox-cloud`
could no longer reach it. The open question was never whether to invert it but
whether inverting would MOVE it in the sequence — it has to run after the
declared seeds land and needs `hostWorkspace`. So it got its own hook,
`afterDeclaredSeeds`, called at exactly the point the by-name call sat, rather
than being folded into `afterSeed` and running earlier. The existing order test
proves the position: mutating the hook earlier in the sequence fails it.

One cycle introduced and backed out, worth recording: `sandbox-cloud`'s test
imported the real stagers from the agent packages, which turbo flagged as
`Circular package dependency: agent-codex, agent-claude, sandbox-cloud`. The
seeding test needs a real tarball, not claude's particular hook filtering, so it
stages with `sandbox-core`'s primitives instead; codex's Keychain-landmine case
— which IS about codex's stager — moved into `agent-codex`'s own tests.

Verified against the real host home: the same five stages dispatch, and claude's
hook-filtered `settings.json` is 31,463 bytes, byte-identical to before the move.
