+++ packages/agent-claude/src/hooks-filter.ts
