+++ packages/agent-codex/package.json
@@ -35,7 +35,8 @@
     "@agentbox/sandbox-cloud": "workspace:*",
     "@agentbox/sandbox-core": "workspace:*",
     "@agentbox/sandbox-docker": "workspace:*",
-    "execa": "^9.5.2"
+    "execa": "^9.5.2",
+    "smol-toml": "^1.7.0"
   },
   "devDependencies": {
     "@types/node": "^22.10.1",
